In this chapter we are going to write a few scripts that simplify the management of our database. Before we get into that let's review how a Python script is executed on the command line.
This was a long tutorial. We have learned the basics of working with a
database, but we haven't incorporated the database into our application
yet. In the next chapter we will put all we have learned about databases
into practice when we look at our user login system.

we created our database and learned how to populate it with users and
posts, but we haven't hooked up any of that into our app yet. And two
chapters ago we've seen how to create web forms and left with a fully
implemented login form.