Configuration
As in previous chapters, we start by configuring the Flask extensions
that we will use. For the login system we will use two extensions, Flask-
Login and Flask-OpenID. These are configured as follows (file
app\__init__.py)

The Flask-OpenID extension requires a path to a temp folder where files
can be stored. For this we provide the location of our tmp folder.