User loader callback
Now we are ready to start implementing the login system using the Flask-
Login and Flask-OpenID extensions.

First, we have to write a function that loads a user from the database.
This function will be used by Flask-Login (file app/views.py)...

Remember that user ids in Flask-Login are always unicode strings, so a
conversion to an integer is necessary before we can send the id to Flask-
SQLAlchemy.