The login view function
Next let's update our login view function (file app/views.py)...

Notice we have imported several new modules, some of which we will use
later.

The changes from our previous version are very small. We have added a
new decorator to our view function. The oid.loginhandler tells Flask-
OpenID that this is our login view function.

At the top of the function body we check if g.user is set to an
authenticated user, and in that case we redirect to the index page. The
idea here is that if there is a logged in user already we will not do a
second login on top.

The g global is setup by Flask as a place to store and share data during
the life of a request. As I'm sure you guessed by now, we will be
storing the logged in user here.

The url_for function that we used in the redirect call is defined by
Flask as a clean way to obtain the URL for a given view function. If you
want to redirect to the index page you may very well use
redirect('/index'), but there are very good reasons to let Flask build
URLs for you.

The code that runs when we get a data back from the login form is also
new. Here we do two things. First we store the value of the remember_me
boolean in the flask session, not to be confused with the db.session
from Flask-SQLAlchemy. We've seen that the flask.g object stores and
shares data though the life of a request. The flask.session provides a
much more complex service along those lines. Once data is stored in the
session object it will be available during that request and any future
requests made by the same client. Data remains in the session until
explicitly removed. To be able to do this, Flask keeps a different
session file for each client of our application.

The oid.try_login call in the following line is the call that triggers
the user authentication through Flask-OpenID. The function takes two
arguments, the openid given by the user in the web form and a list of
data items that we want from the OpenID provider. Since we defined our
User class to include nickname and email, those are the items we are
going to ask for.

The OpenID authentication happens asynchronously. Flask-OpenID will call
a function that is registered with the oid.after_login decorator if the
authentication is successful. If the authentication fails the user will