The Flask-OpenID login callback
Here is our implementation of the after_login function (file
app/views.py)...

The resp argument passed to the after_login function contains
information returned by the OpenID provider.

The first if statement is just for validation. We require a valid email,
so if an email was not provided we cannot log the user in.

Next, we search our database for the email provided. If the email is not
found we consider this a new user, so we add a new user to our database,
pretty much as we have learned in the previous chapter. Note that we
handle the case of a missing nickname, since some OpenID providers may
not have that information.

After that we load the remember_me value from the Flask session, this is
the boolean that we stored in the login view function, if it is
available.

Then we call Flask-Login's login_user function, to register this is a
valid login.

Finally, in the last line we redirect to the next page, or the index
page if a next page was not provided in the request.

The concept of the next page is simple. Let's say you navigate to a page
that requires you to be logged in, but you aren't just yet. In Flask-
Login you can protect views against non logged in users by adding the
login_required decorator. If the user tries to access one of the
affected URLs then it will be redirected to the login page
automatically. Flask-Login will store the original URL as the next page,
and it is up to us to return the user to this page once the login
process completed.