The index view
In a previous chapter we left our index view function using fake
objects, because at the time we did not have users or posts in our
system. Well, we have users now, so let's hook that up:
```
@app.route('/')
@app.route('/index')
@login_required
def index():
    user = g.user
    posts = [
        {
            'author': { 'nickname': 'John' },
            'body': 'Beautiful day in Portland!'
        },
        {
            'author': { 'nickname': 'Susan' },
            'body': 'The Avengers movie was so cool!'
        }
    ]
    return render_template('index.html',
        title = 'Home',
        user = user,
        posts = posts)
```
There are only two changes to this function. First, we have added the
login_required decorator. This will ensure that this page is only seen
by logged in users.

The other change is that we pass g.user down to the template, instead of
the fake object we used in the past.

This is a good time to run the application.

When you navigate to http://localhost:5000 you will instead get the
login page. Keep in mind that to login with OpenID you have to use the
OpenID URL from your provider. You can use one of the OpenID provider
links below the URL text field to generate the correct URL for you.

As part of the login process you will be redirected to your provider's
web site, where you will authenticate and authorize the sharing of some
information with our application (just the email and nickname that we
requested, no passwords or other personal information will be exposed).

Once the login is complete you will be taken to the index page, this
time as a logged in user.

Feel free to try the remember_me checkbox. With this option enabled you
can close and reopen your web browser and will continue to be logged in.