Logging out
We have implemented the log in, now it's time to add the log out.

The view function for logging out is extremely simple (file
app/views.py):
```
@app.route('/logout')
def logout():
    logout_user()
    return redirect(url_for('index'))
```
But we are also missing a link to logout in the template. We are going
to put this link in the top navigation bar which is in the base layout
(file app/templates/base.html):
```
<html>
  <head>
    {% if title %}
    <title>{{title}} - microblog</title>
    {% else %}
    <title>microblog</title>
    {% endif %}
  </head>
  <body>
    <div>Microblog:
        <a href="{{ url_for('index') }}">Home</a>
        {% if g.user.is_authenticated() %}
        | <a href="{{ url_for('logout') }}">Logout</a>
        {% endif %}
    </div>
    <hr>
    {% with messages = get_flashed_messages() %}
    {% if messages %}
    <ul>
    {% for message in messages %}
        <li>{{ message }} </li>
    {% endfor %}
    </ul>
    {% endif %}
    {% endwith %}
    {% block content %}{% endblock %}
  </body>
</html>
```
Note how easy it is to do this. We just needed to check if we have a
valid user set in g.user and if we do we just add the logout link. We
have also used the opportunity to use url_for in our template.