User Profile
Creating a user profile page does not really require any new major
concepts to be introduced. We just need to create a new view function
and its accompanying HTML template.

Here is the view function (file app/views.py):

```
@app.route('/user/<nickname>')
@login_required
def user(nickname):
    user = User.query.filter_by(nickname = nickname).first()
    if user == None:
        flash('User ' + nickname + ' not found.')
        return redirect(url_for('index'))
    posts = [
        { 'author': user, 'body': 'Test post #1' },
        { 'author': user, 'body': 'Test post #2' }
    ]
    return render_template('user.html',
        user = user,
        posts = posts)
```
The @app.route decorator that we used to declare this view function
looks a little bit different than the previous ones. In this case we
have an argument in it, which is indicated as <nickname>. This
translates into an argument of the same name added to the view function.
When the client requests, say, URL /user/miguel the view function will
be invoked with nickname = 'miguel'.

The implementation of the view function should have no surprises. First
we try to load the user from the database, using the nickname that we
received as argument. If that doesn't work then we just redirect to the
main page with an error message, as we have seen in the previous
chapter.

Once we have our user, we just send it in the render_template call,
along with some fake posts. Note that in the user profile page we will
be displaying only posts by this user, so our fake posts have the author
field correctly set.

Our initial view template will be extremely simple (file
app/templates/user.html):

```
<!-- extend base layout -->
{% extends "base.html" %}

{% block content %}
<h1>User: {{user.nickname}}!</h1>
<hr>
{% for post in posts %}
<p>
  {{post.author.nickname}} says: <b>{{post.body}}</b>
</p>
{% endfor %}
{% endblock %}
```
The profile page is now complete, but a link to it does not exist
anywhere in the web site. To make it a bit more easy for a user to check
his or her own profile, we are going to add a link to it in the
navigation bar at the top (file `app/templates/base.html'):
```
    <div>Microblog:
        <a href="{{ url_for('index') }}">Home</a>
        {% if g.user.is_authenticated() %}
        | <a href="{{ url_for('user', nickname = g.user.nickname)
}}">Your Profile</a>
        | <a href="{{ url_for('logout') }}">Logout</a>
        {% endif %}
    </div>
```
Note how in the url_for function we have added the required nickname
argument.

Give the application a try now. Clicking on the Your Profile link at the
top should take you to your user page. Since we don't have any links
that will direct you to an arbitrary user profile page you will need to
type the URL by hand if you want to see someone else. For example, you
would type http://localhost:5000/user/miguel to see the profile of user
miguel.