Avatar
I'm sure you agree that our profile pages are pretty boring. To make
them a bit more interesting, let's add user avatars.

Instead of having to deal with a possibly large collection of uploaded
images in our server, we will rely on the Gravatar service to provide
our user avatars.

Since returning an avatar is a user related task, we will be putting it
in the User class (file app/models.py):
```
from hashlib import md5
# ...
class User(db.Model):
    # ...
    def avatar(self, size):
        return 'http://www.gravatar.com/avatar/' +
md5(self.email).hexdigest() + '?d=mm&s=' + str(size)
```
The avatar method of User returns the URL of the user's avatar image,
scaled to the requested size in pixels.

Turns out with the Gravatar service this is really easy to do. You just
need to create an md5 hash of the user email and then incorporate it
into the specially crafted URL that you see above. After the md5 of the
email you can provide a number of options to customize the avatar. The
d=mm determines what placeholder image is returned when a user does not
have an Gravatar account. The mm option returns the "mystery man" image,
a gray silhouette of a person. The s=N option requests the avatar scaled
to the given size in pixels.

The Gravatar's website has documentation for the avatar URL.

Now that our User class knows how to return avatar images, we can
incorporate them into our profile page layout (file
app/templates/user.html):
```
<!-- extend base layout -->
{% extends "base.html" %}

{% block content %}
<table>
    <tr valign="top">
        <td><img src="{{user.avatar(128)}}"></td>
        <td><h1>User: {{user.nickname}}</h1></td>
    </tr>
</table>
<hr>
{% for post in posts %}
<p>
  {{post.author.nickname}} says: <b>{{post.body}}</b>
</p>
{% endfor %}
{% endblock %}
```
The nice thing about making the User class responsible for returning
avatars is that if some day we decide Gravatar avatars are not what we
want, we just rewrite the avatar method to return different URLs (even
ones that points to our own web server, if we decide we want to host our
own avatars), and all our templates will start showing the new avatars
automatically.

We have added the user avatar to the top of the profile page, but at the
bottom of the page we have posts, and those could have a little avatar
as well. For the user profile page we will of course be showing the same
avatar for all the posts, but then when we move this functionality to
the main page we will have each post decorated with the author's avatar,
and that will look really nice.

To show avatars for the posts we just need to make a small change in our
template (file `app/templates/user.html'):
```
<!-- extend base layout -->
{% extends "base.html" %}

{% block content %}
<table>
    <tr valign="top">
        <td><img src="{{user.avatar(128)}}"></td>
        <td><h1>User: {{user.nickname}}</h1></td>
    </tr>
</table>
<hr>
{% for post in posts %}
<table>
    <tr valign="top">
        <td><img
src="{{post.author.avatar(50)}}"></td><td><i>{{post.author.nickname}}
says:</i><br>{{post.body}}</td>
    </tr>
</table>
{% endfor %}
{% endblock %}
```