Reusing at the sub-template level
We designed the user profile page so that it displays the posts written
by the user. Our index page also displays posts, this time of any user.
So now we have two templates that will need to display posts made by
users. We could just copy/paste the portion of the template that deals
with the rendering of a post, but that is really not ideal, because when
we decide to change the layout of a post we'll have to remember to go
update all the templates that have posts in them.

Instead, we are going to make a sub-template that just renders a post,
then we'll include it in all the templates that need it.

To start, we create a post sub-template, which is really nothing more
than a regular template. We do so by simply extracting the HTML for a
post from our user template (file /app/templates/post.html):
```
<table>
    <tr valign="top">
        <td><img
src="{{post.author.avatar(50)}}"></td><td><i>{{post.author.nickname}}
says:</i><br>{{post.body}}</td>
    </tr>
</table>
```
And then we invoke this sub-template from our user template using
Jinja2's include command (file app/templates/user.html):
```
<!-- extend base layout -->
{% extends "base.html" %}

{% block content %}
<table>
    <tr valign="top">
        <td><img src="{{user.avatar(128)}}"></td>
        <td><h1>User: {{user.nickname}}</h1></td>
    </tr>
</table>
<hr>
{% for post in posts %}
    {% include 'post.html' %}
{% endfor %}
{% endblock %}
```
Once we have a fully functioning index page we will invoke this same sub-
template from over there, but we aren't quite ready to do that, we'll
leave that for a future chapter of this tutorial.