More interesting profiles
More interesting profiles

While we now have a nice profile page, we don't really have much
information to show on it. Users like to tell a bit about them on these
pages, so we'll let them write something about themselves that we can
show here. We will also keep track of what was the last time each user
accessed the site, so that we can also show that on their profile page.

To add these things we have to start by modifying our database. More
specifically, we need to add two new fields to our User class (file
app/models.py):
```
class User(db.Model):
    id = db.Column(db.Integer, primary_key = True)
    nickname = db.Column(db.String(64), unique = True)
    email = db.Column(db.String(120), index = True, unique = True)
    role = db.Column(db.SmallInteger, default = ROLE_USER)
    posts = db.relationship('Post', backref = 'author', lazy =
'dynamic')
    about_me = db.Column(db.String(140))
    last_seen = db.Column(db.DateTime)
```
Every time we modify the database we have to generate a new migration.
Remember that in the database chapter we went through the pain of
setting up a database migration system. We can see the fruits of that
effort now. To add these two new fields to our database we just do this:
```
./db_migrate.py
```
To which our script will respond:

New migration saved as db_repository/versions/003_migration.py
Current database version: 3
And our two new fields are now in our database. Remember that if you are
on Windows the way to invoke this script is different.

If we did not have migration support you would have needed to edit your
database manually, or worse, delete it and recreate it from scratch.

Next, let's modify our profile page template to show these fields (file
app/templates/user.html):
```
<!-- extend base layout -->
{% extends "base.html" %}

{% block content %}
<table>
    <tr valign="top">
        <td><img src="{{user.avatar(128)}}"></td>
        <td>
            <h1>User: {{user.nickname}}</h1>
            {% if user.about_me %}<p>{{user.about_me}}</p>{% endif %}
            {% if user.last_seen %}<p><i>Last seen on:
{{user.last_seen}}</i></p>{% endif %}
        </td>
    </tr>
</table>
<hr>
{% for post in posts %}
    {% include 'post.html' %}
{% endfor %}
{% endblock %}
```
Note we make use of Jinja2's conditionals to show these fields, because
we only want to show them when they are set.

At this point these two new fields are empty for all users, so nothing
will show.

The last_seen field is pretty easy to support. Remember that in a
previous chapter we created a before_request handler, to register the
logged in user with the flask.g global, as g.user. That is the perfect
time to update our database with the last access time for a user (file
app/views.py):
```
from datetime import datetime
# ...
@app.before_request
def before_request():
    g.user = current_user
    if g.user.is_authenticated():
        g.user.last_seen = datetime.utcnow()
        db.session.add(g.user)
        db.session.commit()
```
If you log in to your profile page again the last seen time will now
display, and each time you refresh the page the time will update as
well, because each time the browser makes a request the before_request
handler will update the time in the database.

Note that we are writing the time in the standard UTC timezone. We
discussed this in a previous chapter, we will write all timestamps in
UTC so that they are consistent. That has the undesired side effect that
the time displayed in the user profile page is also in UTC. We will fix
this in a future chapter that will be dedicated to date and time
handling.

To display the user's about me information we have to give them a place
to enter it, and the proper place for this is in the edit profile page.