Let's find a bug
Let's find a bug

I mentioned at the end of the last chapter that I have intentionally
introduced a bug in the application. Let me describe what the bug is,
then we will use it to see what happens to our application when it does
not work as expected.

The problem in the application is that there is no effort to keep the
nicknames of our users unique. The initial nickname of a user is chosen
automatically by the application. If the OpenID provider provides a
nickname for the user then we will just use it. If not we will use the
username part of the email address as nickname. If we get two users with
the same nickname then the second one will not be able to register. To
make matters worse, in the profile edit form we let users change their
nicknames to whatever they want, and again there is no effort to avoid
name collisions.

We will address these problems later, after we analyze how the
application behaves when an error occurs.

Flask debugging support

So let's see what happens when we trigger our bug.

Let's start by creating a brand new database. On Linux:

```
rm app.db
./db_create.py
```
or on Windows:

```
del app.db
flask/Scripts/python db_create.py
```
You need two OpenID accounts to reproduce this bug, ideally from
different providers, so that their cookies don't make this more
complicated. Follow these steps to create a nickname collision:

login with your first account
go to the edit profile page and change the nickname to 'dup'
logout
login with your second account
go to the edit profile page and change the nickname to 'dup'
Oops! We've got an exception from sqlalchemy. The text of the error
reads:

```
sqlalchemy.exc.IntegrityError
IntegrityError: (IntegrityError) column nickname is not unique u'UPDATE
user SET nickname=?, about_me=? WHERE user.id = ?' (u'dup', u'', 2)
```
What follows after the error is a stack trace of the error, and actually
it is a pretty nice one, where you can go to any frame and inspect
source code or even evaluate expressions right in the browser.

The error is pretty clear, we tried to insert a duplicated nickname in
the database. The database model had a unique constrain on the nickname
field, so this is an invalid operation.

In addition to the actual error, we have a secondary problem in our
hands. If a user inadvertently causes an error in our application (this
one or any other that causes an exception) it will be him or her that
gets the error with the revealing error message and the stack trace, not
us. While this is a fantastic feature while we are developing, it is
something we definitely do not want our users to ever see.

All this time we have been running our application in debug mode. The
debug mode is enabled when the application starts, by passing a debug =
True argument to the run method. This is how we coded our run.py start-
up script.

When we are developing the application this is convenient, but we need
to make sure it is turned off when we run our application in production
mode. Let's just create another starter script that runs with debugging
disabled (file runp.py):
```
#!flask/bin/python
from app import app
app.run(debug = False)
```
Now restart the application with:
```
./runp.py
```
And now try again to rename the nickname on the second account to 'dup'.

This time we did not get an error. Instead, we've got an HTTP error code
500, which is Internal Server Error. Not a great looking error, but at
least we are not exposing any details of our application to strangers.
The error 500 page is generated by Flask when debugging is off and an
unhandled exception occurs.

While this is better, we are now having two new issues. First a cosmetic
one: the default error 500 page is ugly. The second problem is much more
important. With things as they are we would never know when and if a
user experiences a failure in our application because now application
failures are silently dismissed. Luckily there are easy ways to address
both problems.