Custom HTTP error handlers
Custom HTTP error handlers

Flask provides a mechanism for an application to install its own error
pages. As an example, let's define custom error pages for the HTTP
errors 404 and 500, the two most common ones. Defining pages for other
errors works in the same way.

To declare a custom error handler the errorhandler decorator is used
(file app/views.py):
```
@app.errorhandler(404)
def internal_error(error):
    return render_template('404.html'), 404

@app.errorhandler(500)
def internal_error(error):
    db.session.rollback()
    return render_template('500.html'), 500
```
Not much to talk about for these, as they are almost self-explanatory.
The only interesting bit is the rollback statement in the error 500
handler. This is necessary because this function will be called as a
result of an exception. If the exception was triggered by a database
error then the database session is going to arrive in an invalid state,
so we have to roll it back in case a working session is needed for the
rendering of the template for the 500 error.

Here is the template for the 404 error:

```
<!-- extend base layout -->
{% extends "base.html" %}

{% block content %}
<h1>File Not Found</h1>
<p><a href="{{url_for('index')}}">Back</a></p>
{% endblock %}
```
And here is the one for the 500 error:

```
<!-- extend base layout -->
{% extends "base.html" %}

{% block content %}
<h1>An unexpected error has occurred</h1>
<p>The administrator has been notified. Sorry for the inconvenience!</p>
<p><a href="{{url_for('index')}}">Back</a></p>
{% endblock %}
```
Note that in both cases we continue to use our base.html layout, so that
the error page has the look and feel of the application.