Logging to a file
Logging to a file

Receiving errors via email is nice, but sometimes this isn't enough.
There are some failure conditions that do not end in an exception and
aren't a major problem, yet we may want to keep track of them in a log
in case we need to do some debugging.

For this reason, we are also going to maintain a log file for the
application.

Enabling file logging is similar to the email logging (file
app/__init__.py):
```
if not app.debug:
    import logging
    from logging.handlers import RotatingFileHandler
    file_handler = RotatingFileHandler('tmp/microblog.log', 'a', 1 *
1024 * 1024, 10)
    file_handler.setFormatter(logging.Formatter('%(asctime)s
%(levelname)s: %(message)s [in %(pathname)s:%(lineno)d]'))
    app.logger.setLevel(logging.INFO)
    file_handler.setLevel(logging.INFO)
    app.logger.addHandler(file_handler)
    app.logger.info('microblog startup')
```
The log file will go to our tmp directory, with name microblog.log. We
are using the RotatingFileHandler so that there is a limit to the amount
of logs that are generated. In this case we are limiting the size of a
log file to one megabyte, and we will keep the last ten log files as
backups.

The logging.Formatter class provides custom formatting for the log
messages. Since these messages are going to a file, we want them to have
as much information as possible, so we write a timestamp, the logging
level and the file and line number where the message originated in
addition to the log message and the stack trace.

To make the logging more useful, we are lowering the logging level, both
in the app logger and the file logger handler, as this will give us the
opportunity to write useful messages to the log without having to call
them errors. As an example, we start by logging the application start up
as an informational level. From now on, each time you start the
application without debugging the log will record the event.

While we don't have a lot of need for a logger at this time, debugging a
web server that is online and in use is very difficult. Logging messages
to a file is an extremely useful tool in diagnosing and locating issues,
so we are now all ready to go should we need to use this feature.