The bug fix
The bug fix

Let's fix our nickname duplication bug.

As discussed earlier, there are two places that are currently not
handling duplicates. The first is in the after_login handler for Flask-
Login. This is called when a user successfully logs in to the system and
we need to create a new User instance. Here is the affected snippet of
code, with the fix in it (file app/views.py):
```
    if user is None:
        nickname = resp.nickname
        if nickname is None or nickname == "":
            nickname = resp.email.split('@')[0]
        nickname = User.make_unique_nickname(nickname)
        user = User(nickname = nickname, email = resp.email, role =
ROLE_USER)
        db.session.add(user)
        db.session.commit()
```
The way we solve the problem is by letting the User class pick a unique
name for us. This is what the new make_unique_nickname method does (file
app/models.py):
```

    class User(db.Model):
    # ...
    @staticmethod
    def make_unique_nickname(nickname):
        if User.query.filter_by(nickname = nickname).first() == None:
            return nickname
        version = 2
        while True:
            new_nickname = nickname + str(version)
            if User.query.filter_by(nickname = new_nickname).first() ==
None:
                break
            version += 1
        return new_nickname
    # ...
```
This method simply adds a counter to the requested nickname until a
unique name is found. For example, if the username "miguel" exists, the
method will suggest "miguel2", but if that also exists it will go to
"miguel3" and so on. Note that we coded the method as a static method,
since it this operation does not apply to any particular instance of the
class.

The second place where we have problems with duplicate nicknames is the
view function for the edit profile page. This one is a little tricker to
handle, because it is the user choosing the nickname. The correct thing
to do here is to not accept a duplicated nickname and let the user enter
another one. We will address this by adding custom validation to the
nickname form field. If the user enters an invalid nickname we'll just
fail the validation for the field, and that will send the user back to
the edit profile page. To add our validation we just override the form's
validate method (file app/forms.py):

```
class EditForm(Form):
    nickname = TextField('nickname', validators = [Required()])
    about_me = TextAreaField('about_me', validators = [Length(min = 0,
max = 140)])

    def __init__(self, original_nickname, *args, **kwargs):
        Form.__init__(self, *args, **kwargs)
        self.original_nickname = original_nickname

    def validate(self):
        if not Form.validate(self):
            return False
        if self.nickname.data == self.original_nickname:
            return True
        user = User.query.filter_by(nickname =
self.nickname.data).first()
        if user != None:
            self.nickname.errors.append('This nickname is already in
use. Please choose another one.')
            return False
        return True
```
The form constructor now takes a new argument original_nickname. The
validate method uses it to determine if the nickname has changed or not.
If it hasn't changed then it accepts it. If it has changed, then it
makes sure the new nickname does not exist in the database.

To complete this change we have to enable field errors to show in our
template for the form (file app/templates/edit.html):
```

        <td>Your nickname:</td>
        <td>
            {{form.nickname(size = 24)}}
            {% for error in form.errors.nickname %}
            <br><span style="color: red;">[{{error}}]</span>
            {% endfor %}
        </td>
```
Now the bug is fixed and duplicates will be prevented... except when
they are not. We still have a potential problem with concurrent access
to the database by two or more threads or processes, but this will be
the topic of a future article.

At this point you can try again to select a duplicated name to see how
the form nicely handles the error.