Adding and removing 'follows'
Adding and removing 'follows'

To promote reusability, we will implement the follow and unfollow
functionality in the User model instead of doing it directly in view
functions. That way we can use this feature for the actual application
(invoking it from the view functions) and also from our unit testing
framework. As a matter of principle, it is always best to move the logic
of our application away from view functions and into models, because
that simplifies the testing. You want to have your view functions be as
simple as possible, because those are harder to test in an automated
way.

Below is the code to add and remove relationships, defined as methods of
the User model (file app/models.py):
```
class User(db.Model):
    #...
    def follow(self, user):
        if not self.is_following(user):
            self.followed.append(user)
            return self

    def unfollow(self, user):
        if self.is_following(user):
            self.followed.remove(user)
            return self

    def is_following(self, user):
        return self.followed.filter(followers.c.followed_id ==
user.id).count() > 0
```
These methods are amazingly simple, thanks to the power of sqlalchemy
who does a lot of work under the covers. We just add or remove items
from the followed relationship and sqlalchemy takes care of managing the
association table for us.

The follow and unfollow methods are defined so that they return an
object when they succeed or None when they fail. When an object is
returned, this object has to be added to the database session and
committed.

The is_following method does a lot in its single line of code. We are
taking the followed relationship query, which returns all the (follower,
followed) pairs that have our user as the follower, and we filter it by
the followed user. This is possible because the followed relationship
has a lazy mode of dynamic, so instead of being the result of the query,
this is the actual query object, before execution.

The return from the filter call is the modified query, still without
having executed. So we then call count() on this query, and now the
query will execute and return the number of records found. If we get
one, then we know a link between these two uses is already present. If
we get none then we know a link does not exist.