Being your own follower
Tying up loose ends

We have made an impressive amount of progress today. But while we have
solved all the problems related to database setup and querying, we have
not enabled the new functionality through our application.

Luckily for us, there aren't any challenges in doing this. We just need
to fix view functions and templates to call the new methods in the User
model when appropriate. So let's do that before we close this session.

Being your own follower

We decided that we were going to mark all users as followers of
themselves, so that they can see their own posts in their post stream.

We are going to do that at the point where users are getting their
accounts setup, in the after_login handler for OpenID (file
'app/views.py'):

@oid.after_login
def after_login(resp):
    if resp.email is None or resp.email == "":
        flash('Invalid login. Please try again.')
        redirect(url_for('login'))
    user = User.query.filter_by(email = resp.email).first()
    if user is None:
        nickname = resp.nickname
        if nickname is None or nickname == "":
            nickname = resp.email.split('@')[0]
        user = User(nickname = nickname, email = resp.email, role =
ROLE_USER)
        db.session.add(user)
        db.session.commit()
        # make the user follow him/herself
        db.session.add(user.follow(user))
        db.session.commit()
    remember_me = False
    if 'remember_me' in session:
        remember_me = session['remember_me']
        session.pop('remember_me', None)
    login_user(user, remember = remember_me)
    return redirect(request.args.get('next') or url_for('index'))